Allow to ignore some labels
Prometheus is sensitive to cardinality of metrics and can blow up if it
becomes too high. The thing is that each possible value of a label will
double the cardinality of the endpoint.

A solution is too remove some labels for a given metric. This patch
allows to set labels to be ignored for a given metric.